refactor: Dashboard 페이지 import 리팩토링
